iota_snapshot::restore

Trait Restore

pub trait Restore {
    // Required method
    fn insert_partition(
        &self,
        file_metadata: FileMetadata,
        bytes: Bytes,
        expected_checksum: &[u8; 32],
    ) -> impl Future<Output = Result<()>> + Send;
}

A trait for databases that can be restored from a formal snapshot.

Required Methods

fn insert_partition( &self, file_metadata: FileMetadata, bytes: Bytes, expected_checksum: &[u8; 32], ) -> impl Future<Output = Result<()>> + Send

Inserts a partition of live objects.

The checksum of the partition can be computed as the SHA3 hash of all objects’ digests included. Then it can be verified against the given expected value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.